+++ server/component/style/openAppButton/OpenAppButtonView.php
@@ -1,0 +1,43 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<?php
+require_once __DIR__ . "/../../../../../../component/style/StyleView.php";
+
+/**
+ * The view class of the card style component.
+ * 
+ */
+class OpenAppButtonView extends StyleView
+{
+    /* Private Properties *****************************************************/
+
+    /* Constructors ***********************************************************/
+
+    /**
+     * The constructor.
+     *
+     * @param object $model
+     *  The model instance of a base style component.
+     */
+    public function __construct($model)
+    {
+        parent::__construct($model);
+    }
+
+    /* Private Methods ********************************************************/
+
+    /* Public Methods *********************************************************/
+
+    /**
+     * Render the style view.
+     */
+    public function output_content()
+    {
+        require __DIR__ . "/tpl_openAppButton.php";
+    }
+
+}
+?>